+++ src/app/valentina/mainwindow.cpp
@@ -3707,7 +3707,7 @@ void MainWindow::DoExport(const VCommandLinePtr &expParams)
         {
             DialogSaveLayout dialog(scenes.size(), expParams->OptBaseName(), this);
             dialog.SetDestinationPath(expParams->OptDestinationPath());
-            dialog.SelectFormate(expParams->OptExportType());
+            dialog.SelectFormate(static_cast<quint32>(expParams->OptExportType()));
             ExportLayout(dialog);
         }
         catch (const VException &e)
